requirements.txt
Flask==2.3.3
flask-cors==6.0.1
gunicorn==23.0.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
requests==2.31.0
tensorflow==2.13.0
tensorflow-hub==0.14.0
